gtk stuff

+++ snd-g1.h
@@ -169,7 +169,7 @@ void set_menu_label(GtkWidget *w, const char *label);
 int g_change_menu_label(int which_menu, char *old_label, char *new_label);
 int g_set_menu_sensitive(int which_menu, char *old_label, int on);
 int g_menu_is_sensitive(int which_menu, char *old_label);
-int g_add_to_main_menu(snd_state *ss, char *label, long slot);
+int g_add_to_main_menu(snd_state *ss, char *label, int slot);
 int g_add_to_menu(snd_state *ss, int which_menu, char *label, int callb, int position);
 int g_remove_from_menu(int which_menu, char *label);
 GtkWidget *add_menu(snd_state *state);
